# crAPIEncryption/api-specs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OWASP crAPI API",
    "version": "1.0"
  },
...
  "paths": {
...
    "/identity/api/v2/vehicle/vehicles": {
      "get": {
        "operationId": "OP_IDENT_VEHICLE_GET_VEHICLES",
        "summary": "Get user vehicles",
        "tags": [
          "Identity / Vehicle"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_ADMIN"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "model",
                      "owner",
                      "pincode",
                      "status",
                      "uuid",
                      "vehicleLocation",
                      "vin",
                      "year"
                    ],
                    "properties": {
                      "model": {
                        "type": "object",
                        "required": [
                          "fuel_type",
                          "id",
                          "model",
                          "vehicle_img",
                          "vehiclecompany"
                        ],
                        "properties": {
                          "model": {
                            "type": "string"
                          },
                          "fuel_type": {
                            "type": "string"
                          },
                          "vehicle_img": {
                            "type": "string"
                          },
                          "id": {
                            "type": "number"
                          },
                          "vehiclecompany": {
                            "type": "object",
                            "required": [
                              "id",
                              "name"
                            ],
                            "properties": {
                              "id": {
                                "type": "number"
                              },
                              "name": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      },
                      "uuid": {
                        "type": "string"
                      },
                      "vehicleLocation": {
                        "type": "object",
                        "required": [
                          "id",
                          "latitude",
                          "longitude"
                        ],
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "latitude": {
                            "type": "string"
                          },
                          "longitude": {
                            "type": "string"
                          }
                        }
                      },
                      "year": {
                        "type": "number"
                      },
                      "id": {
                        "type": "number"
                      },
                      "status": {
                        "type": "string"
                      },
                      "vin": {
                        "type": "string"
                      },
                      "owner": {},
                      "pincode": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/identity/api/v2/vehicle/add_vehicle": {
      "post": {
        "operationId": "OP_IDENT_VEHICLE_ADD_VEHICLE",
        "summary": "Add the user vehicle",
        "tags": [
          "Identity / Vehicle"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_ADMIN"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pincode",
                  "vin"
                ],
                "properties": {
                  "pincode": {
                    "type": "string"
                  },
                  "vin": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "status"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "status"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
    "/identity/api/v2/vehicle/{vehicleId}/location": {
      "get": {
        "operationId": "OP_IDENT_VEHICLE_GET_LOCATION",
        "summary": "Get user's vehicle location",
        "tags": [
          "Identity / Vehicle"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_ADMIN"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "carId",
                    "fullName",
                    "vehicleLocation"
                  ],
                  "properties": {
                    "carId": {
                      "type": "string"
                    },
                    "fullName": {
                      "type": "string"
                    },
                    "vehicleLocation": {
                      "type": "object",
                      "required": [
                        "id",
                        "latitude",
                        "longitude"
                      ],
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "latitude": {
                          "type": "string"
                        },
                        "longitude": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "status"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": [
        {
          "in": "path",
          "name": "vehicleId",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ]
    },
    "/identity/api/v2/vehicle/resend_email": {
      "post": {
        "operationId": "OP_IDENT_VEHICLE_RESEND_EMAIL",
        "summary": "Resend vehicles details to be added to the user dashboard",
        "tags": [
          "Identity / Vehicle"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-levo-allowed-roles": [
          "ROLE_USER",
          "ROLE_ADMIN"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "status"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "description": ""
          }
        }
      },
      "parameters": []
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}